+++ public/application-assets/gulp/01-clean.js
@@ -2,7 +2,8 @@ var gulp = require('gulp');
 var rimraf = require('rimraf').sync;
 
 // Erases the dist folder
-gulp.task('clean', function() {
+gulp.task('clean', function(cb) {
   rimraf('_build');
   rimraf('dist');
+  cb();
 });