+++ public/application-assets/gulpfile.js
@@ -1,7 +1,6 @@
 var gulp = require('gulp');
 var browser = require('browser-sync');
 var requireDir = require('require-dir');
-var sequence = require('run-sequence');
 var exec = require('child_process').execSync;
 var port = process.env.SERVER_PORT || 3000;
 
@@ -27,35 +26,29 @@ gulp.task('devcopy', function(cb) {
 
 
 // Builds the documentation and framework files
-gulp.task('prepare', function (cb) {
-	sequence('clean', 'copy', 'docs:all', cb);
-});
+gulp.task('prepare', gulp.series('clean', 'copy'/*, function (done) { done(); }*/));
 
 // Builds the documentation and framework files
 gulp.task('completed', function (cb) {
   console.log('build completed at ' + (new Date).toString() + '...');
-  //cb();
+  cb();
 });
 
-gulp.task('build', function (cb) {
-	sequence('sass', 'javascript', 'deploy', function () {
-    console.log('build completed at ' + (new Date).toString() + '...');
-    cb();
-  });
-});
+gulp.task('build', gulp.series('sass', 'javascript', /*'docs:all',*/ 'deploy', function (cb) {
+  console.log('build completed at ' + (new Date).toString() + '...');
+  cb();
+}));
 
 // Starts a BrowerSync instance
-gulp.task('serve', ['prepare', 'build'], function(){
+gulp.task('serve', gulp.series('prepare', 'build', function(cb){
   console.log('serving...');
   browser.init({server: './_build', port: port});
-});
-
-//Runs all of the above tasks and then waits for files to change
-gulp.task('default', ['prepare', 'build', 'watch']);
+  cb();
+}));
 
 gulp.task('watch', function() {
   console.log('watching...');
-  gulp.watch('src/**/*', ['build']);
+  gulp.watch(['src/**/*', '!src/scss/settings/_settings.scss'], gulp.series('build'));
   //gulp.watch('src/docs/**/*', ['docs', browser.reload]);
   //gulp.watch(['src/docs/layout/*.html', 'src/docs/partials/*.html'], ['docs:all', browser.reload]);
   //gulp.watch('src/scss/**/*', ['sass', browser.reload]);
@@ -66,3 +59,7 @@ gulp.task('watch', function() {
 
 // Runs all of the above tasks and then waits for files to change
 //gulp.task('default', ['build', 'watch']);
+
+
+//Runs all of the above tasks and then waits for files to change
+gulp.task('default', gulp.series('prepare', 'build', 'watch'));